+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## [3.0.3] - 2020.5.16
+
+* Add `minLength` option to not use `compute` when text is small.
+* Disable `compute` on web. Can be overriden using `disableOnWeb`.
+
 ## [3.0.2] - 2019.9.19
 
 * A dio transformer especially for flutter, by which the json decoding will be in background with [compute] function.